refactor(daemon): dispatch RPC through a handler registry
`JsonRpcServer.handleRequest` dispatched 24 request methods from one
`when (req.method)`, so every feature that added a method grew this one
file and every feature's handler lived far from the feature it serves.
`daemon/core` already solves that shape elsewhere — `ExtensionRegistry`,
`RecordingScriptHandlerRegistry`, `DataProductRegistry`,
`FrameStreamRegistry` — but RPC dispatch never got the same treatment.

Add `RpcMethodRegistry`: a method-name → handler map built once at
construction, plus a narrow `RpcPeer` port carrying the four things a
handler needs from the connection (decode params, encode a result, send
a response, send an error). `handleRequest` becomes a lookup and a call;
an unregistered method is what `method not found` means.

Take the first slice the issue names: the history cluster moves to
`history/HistoryRpcHandlers`, beside the `HistoryManager` it dispatches
into, lifting ~470 lines out of `JsonRpcServer` (4675 to 4215). The
handler bodies are unchanged apart from reaching the connection through
`RpcPeer`; the two feature gates (the 1.1 history flag, the
`history/diff` experiment) move into handlers that reply with the gate's
error, so a gated method is still registered and the absent case stays
unambiguous.

`RpcPeer` is an adapter object rather than `JsonRpcServer : RpcPeer`, so
the reply helpers stay private and the server's public API is unchanged.

Refs #5166
