+++ docs/AGENT_GUIDE.md
@@ -169,6 +169,7 @@ VS Code extension ([`src/`](https://github.com/yschimke/compose-preview-vscode/b
 
 Daemon (`daemon/core/src/main/kotlin/ee/schimke/composeai/daemon/`):
 
+- **`RpcMethodRegistry`** ([rpc/RpcMethodRegistry.kt](../daemon/core/src/main/kotlin/ee/schimke/composeai/daemon/rpc/RpcMethodRegistry.kt)) — method-name → handler map for the JSON-RPC request surface. `JsonRpcServer.handleRequest` is a lookup plus a call, not a `when (req.method)`; a new method registers a handler where its feature lives (`history/*` is served by `history/HistoryRpcHandlers`, which reaches the connection through the narrow `RpcPeer` port). **Don't add an arm to a central dispatch table** — an unregistered method is what `method not found` means, so a gated method registers a handler that replies with the gate's error rather than being omitted.
 - **`SubscriptionStore`** ([SubscriptionStore.kt](../daemon/core/src/main/kotlin/ee/schimke/composeai/daemon/SubscriptionStore.kt)) — sticky `(previewId, kind)` bookkeeping. Three teardown methods (`unsubscribe`, `retainVisible`, `removeKinds`) return the dropped pairs so the caller routes `onUnsubscribe` through the right producer surface (`publicDataProducts` vs `activeDataProducts`).
 - **`DeferredDiscoveryQueue`** ([DeferredDiscoveryQueue.kt](../daemon/core/src/main/kotlin/ee/schimke/composeai/daemon/DeferredDiscoveryQueue.kt)) — paths waiting for the post-render discovery cascade plus the watchdog that drains them if no render arrives. Watchdog scheduler is injectable; tests use a manual scheduler so race outcomes are deterministic.
 - **`DataProductRegistry.renderModeFor(kind)`** ([DataProductRegistry.kt](../daemon/core/src/main/kotlin/ee/schimke/composeai/daemon/DataProductRegistry.kt)) — producers declare which kinds need a renderer mode tag (`"a11y"`, etc.). The dispatcher iterates subscribed kinds and asks the registry; **do not pattern-match on the kind string in `JsonRpcServer`**.
